ApplymentService: Change return type to ResponseEntity

+++ src/main/java/com/pyunggang/churchproject/service/impl/ApplymentServiceImpl.java
@@ -2,12 +2,13 @@ package com.pyunggang.churchproject.service.impl;
 
 import com.pyunggang.churchproject.data.dto.ApplymentParam;
 import com.pyunggang.churchproject.data.entity.Applyment;
-import com.pyunggang.churchproject.data.entity.Event;
 import com.pyunggang.churchproject.data.entity.Participant;
 import com.pyunggang.churchproject.data.repository.*;
 import com.pyunggang.churchproject.service.ApplymentService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,9 +34,10 @@ public class ApplymentServiceImpl implements ApplymentService {
 //    TODO: 동명이인 어떡할지 고민 필요
 //    TODO: 런타임 에러 발생시켜서 롤백 시킬 방법 찾기
     @Override
-    public boolean saveApplyment(List<ApplymentParam> params) {
+    public ResponseEntity saveApplyment(List<ApplymentParam> params) {
         Participant participant;
         Applyment applyment;
+        HttpStatus status = HttpStatus.OK;
 
         for (ApplymentParam param : params) {
             // 빈 데이터가 있는 참가자 정보 건너뛰기
@@ -56,15 +58,16 @@ public class ApplymentServiceImpl implements ApplymentService {
             // 이미 등록된 참가자가 아니라면,
             if (!partiRepo.existsByParticipant(participant)) {
                 // Participant에 삽입 / 실패하면 false 리턴
-                if (!partiRepo.save(participant).getName().equals(param.getName()))
-                    return false;
+                if (!partiRepo.save(participant).getName().equals(param.getName())) {
+                    status = HttpStatus.INTERNAL_SERVER_ERROR;
+                    break;
+                }
             } else {
                 // 이미 등록된 참가자라면,
                 // 중복 신청인지, 다른 종목에 참여하는 같은 참가자인지 확인
-                // 중복 신청이라면, return false
-                if (applyRepo.existsByEventNameAndParticipant(param.getEventName(), participant)) {
-                    return false;
-                }
+                // 중복 신청이라면 건너뜀
+                if (applyRepo.existsByEventNameAndParticipant(param.getEventName(), participant))
+                    continue;
 
                 // 저장되어 있는 Participant를 불러온다
                 participant = partiRepo.findParticipantByParticipant(participant);
@@ -79,11 +82,13 @@ public class ApplymentServiceImpl implements ApplymentService {
                     .build();
 
             // Applyment에 삽입 / 실패하면 return false
-            if (!applyRepo.save(applyment).getParticipant().getName().equals(param.getName()))
-                return false;
+            if (!applyRepo.save(applyment).getParticipant().getName().equals(param.getName())) {
+                status = HttpStatus.INTERNAL_SERVER_ERROR;
+                break;
+            }
         }
 
-        return true;
+        return new ResponseEntity(status);
     }
 
     /**
@@ -93,14 +98,21 @@ public class ApplymentServiceImpl implements ApplymentService {
      * @return ParticipantRegisterParam - List 형태로 리턴
      */
     @Override
-    public List<ApplymentParam> findApplymentList(String churchName, String eventName) {
+    public ResponseEntity<List<ApplymentParam>> findApplymentList(String churchName, String eventName) {
+        HttpStatus status = HttpStatus.OK;
         // 빈 linked list 생성
         List<ApplymentParam> params = new LinkedList<>();
-        // 교회명과 종목명으로 불러온 applyment를 하나하나 ParticipantRegisterParam으로 변경
-        for (Applyment applyment : applyRepo.findAllByEventNameAndParticipantChurchName(eventName, churchName)) {
-            params.add(new ApplymentParam(applyment));
+        List<Applyment> applyments = applyRepo.findAllByEventNameAndParticipantChurchName(eventName, churchName).orElseGet(null);
+
+        if (applyments == null)
+            status = HttpStatus.NOT_FOUND;
+        else {
+            for (Applyment applyment : applyments) {
+                params.add(new ApplymentParam(applyment));
+            }
         }
-        return params;
+
+        return new ResponseEntity<>(params, status);
     }
 
     /**
@@ -113,35 +125,47 @@ public class ApplymentServiceImpl implements ApplymentService {
      */
     @Override
     @Transactional
-    public boolean deleteApplyment(String eventName, int participantId) {
+    public ResponseEntity deleteApplyment(String eventName, int participantId) {
+        HttpStatus status = HttpStatus.OK;
         // id로 참가자 정보 불러오기
-        Participant participant = partiRepo.findById(participantId).get();
+        Participant participant = partiRepo.findById(participantId).orElseGet(null);
         // applyment 정보 불러오기
-        Applyment applyment = applyRepo.findByParticipantAndEventName(participant, eventName);
-        // 신청정보 삭제
-        applyRepo.delete(applyment);
+        Applyment applyment = applyRepo.findByParticipantAndEventName(participant, eventName).orElse(null);
 
-        // 같은 참가자가 다른 종목을 신청하지 않았으면, 참가자 정보 삭제
-        if (!applyRepo.existsByParticipant(participant)) {
-            partiRepo.delete(participant);
+        if (participant == null || applyment == null) {
+            status = HttpStatus.BAD_REQUEST;
+        } else {
+            // 신청정보 삭제
+            applyRepo.delete(applyment);
+
+            // 같은 참가자가 다른 종목을 신청하지 않았으면, 참가자 정보 삭제
+            if (!applyRepo.existsByParticipant(participant)) {
+                partiRepo.delete(participant);
+            }
         }
 
-        return true;
+        return new ResponseEntity(status);
     }
 
     @Override
     @Transactional
-    public boolean updateApplyment(ApplymentParam applymentParam) {
-        Participant participant = partiRepo.findById(applymentParam.getId()).get();
+    public ResponseEntity updateApplyment(ApplymentParam applymentParam) {
+        HttpStatus status = HttpStatus.CREATED;
+        Participant participant = partiRepo.findById(applymentParam.getId()).orElseGet(null);
 
-        participant.setAge(applymentParam.getAge());
-        participant.setName(applymentParam.getName());
-        participant.setGender(applymentParam.getGender());
-        participant.setGrade(applymentParam.getGrade());
-        participant.setDepartment(departRepo.findDepartmentByNameIs(applymentParam.getDepartment()));
+        if (participant == null)
+            status = HttpStatus.BAD_REQUEST;
+        else {
+            participant.setAge(applymentParam.getAge());
+            participant.setName(applymentParam.getName());
+            participant.setGender(applymentParam.getGender());
+            participant.setGrade(applymentParam.getGrade());
+            participant.setDepartment(departRepo.findDepartmentByNameIs(applymentParam.getDepartment()));
 
-        partiRepo.save(participant);
+            if (partiRepo.save(participant).getParticipantId() != participant.getParticipantId())
+                status = HttpStatus.INTERNAL_SERVER_ERROR;
+        }
 
-        return true;
+        return new ResponseEntity(status);
     }
 }
